refactor: NaverBookItem 매퍼 함수 위치 변경
